+++ android/src/main/kotlin/hamza/dali/flutter_osm_plugin/FlutterOsmView.kt
@@ -31,6 +31,9 @@ import hamza.dali.flutter_osm_plugin.FlutterOsmPlugin.Companion.PAUSED
 import hamza.dali.flutter_osm_plugin.FlutterOsmPlugin.Companion.STARTED
 import hamza.dali.flutter_osm_plugin.FlutterOsmPlugin.Companion.STOPPED
 import hamza.dali.flutter_osm_plugin.FlutterOsmPlugin.Companion.mapSnapShots
+import hamza.dali.flutter_osm_plugin.models.FlutterMarker
+import hamza.dali.flutter_osm_plugin.models.FlutterRoad
+import hamza.dali.flutter_osm_plugin.models.RoadConfig
 import hamza.dali.flutter_osm_plugin.utilities.*
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding.OnSaveInstanceStateListener
 import io.flutter.plugin.common.BinaryMessenger
@@ -39,12 +42,10 @@ import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.PluginRegistry
 import io.flutter.plugin.platform.PlatformView
-import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.*
 import kotlinx.coroutines.Dispatchers.Default
+import kotlinx.coroutines.Dispatchers.IO
 import kotlinx.coroutines.Dispatchers.Main
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import org.osmdroid.bonuspack.routing.OSRMRoadManager
 import org.osmdroid.bonuspack.routing.RoadManager
 import org.osmdroid.bonuspack.utils.PolylineEncoder
@@ -440,6 +441,12 @@ class FlutterOsmView(
                 "road" -> {
                     drawRoad(call, result)
                 }
+                "draw#multi#road" -> {
+                    drawMultiRoad(call, result)
+                }
+                "clear#roads" -> {
+                    clearAllRoad(result)
+                }
                 "marker#icon" -> {
                     changeIcon(call, result)
                 }
@@ -606,7 +613,7 @@ class FlutterOsmView(
                 val polyLine = Polyline(map!!)
                 polyLine.setPoints(lastRoad.roadPoints)
                 //customRoadMarkerIcon.p
-                createRoad(
+                flutterRoad = createRoad(
                     polyLine = polyLine,
                     colorRoad = lastRoad.roadColor,
                     roadWidth = lastRoad.roadWith,
@@ -618,7 +625,24 @@ class FlutterOsmView(
             }
 
         }
-
+        mapSnapShot.cachedRoads().forEach { road ->
+            if (road.roadPoints.isNotEmpty()) {
+                if (!map!!.overlayManager.contains(folderRoad)) {
+                    map!!.overlayManager.add(folderRoad)
+                }
+                val polyLine = Polyline(map!!)
+                polyLine.setPoints(road.roadPoints)
+                //customRoadMarkerIcon.p
+                flutterRoad = createRoad(
+                    polyLine = polyLine,
+                    colorRoad = road.roadColor,
+                    roadWidth = road.roadWith,
+                    listInterestPoints = road.listInterestPoints,
+                    showPoiMarker = road.showIcons,
+                )
+            }
+        }
+        map!!.invalidate()
         resetAdvPickerOrTrackLocation(mapSnapShot)
         clearCacheMap()
         methodChannel.invokeMethod("map#restored", null)
@@ -920,7 +944,7 @@ class FlutterOsmView(
             animateTo = false
         )
 
-        //map!!.overlays.add(marker)
+
         result.success(null)
 
     }
@@ -1018,12 +1042,11 @@ class FlutterOsmView(
         val colorRoad = (args["roadColor"] as List<Int>)
         val color = Color.rgb(colorRoad.first(), colorRoad.last(), colorRoad[1])
         val widthRoad = (args["roadWidth"] as Double)
+        val zoomToRegion = args["zoomInto"] as Boolean
+        checkRoadFolderAboveUserOverlay()
 
-        if (!map!!.overlays.contains(folderRoad)) {
-            map!!.overlays.add(folderRoad)
-        } else {
-            folderRoad.items.clear()
-        }
+        folderRoad.items.clear()
+
 
         val route = PolylineEncoder.decode(encodedWayPoints, 10, false)
         val polyLine = Polyline(map!!)
@@ -1041,7 +1064,13 @@ class FlutterOsmView(
                 showIcons = false,
             )
         )
-
+        if (zoomToRegion) {
+            map!!.zoomToBoundingBox(
+                BoundingBox.fromGeoPoints(polyLine.actualPoints),
+                true,
+                64,
+            )
+        }
         map!!.invalidate()
         result.success(null)
     }
@@ -1302,6 +1331,140 @@ class FlutterOsmView(
         result.success(null)
     }
 
+    private fun clearAllRoad(result: MethodChannel.Result) {
+        folderRoad.items.clear()
+
+        map!!.invalidate()
+        result.success(200)
+    }
+
+    private fun drawMultiRoad(call: MethodCall, result: MethodChannel.Result) {
+        val args = call.arguments!! as List<HashMap<String, Any>>
+        val listConfigRoad = emptyList<RoadConfig>().toMutableList()
+
+        for (arg in args) {
+            val waypoints = (arg["wayPoints"] as List<HashMap<String, Double>>).map { map ->
+                GeoPoint(map["lat"]!!, map["lon"]!!)
+            }.toList()
+            listConfigRoad.add(
+                RoadConfig(
+                    meanUrl = when (arg["roadType"] as String) {
+                        "car" -> OSRMRoadManager.MEAN_BY_CAR
+                        "bike" -> OSRMRoadManager.MEAN_BY_BIKE
+                        "foot" -> OSRMRoadManager.MEAN_BY_FOOT
+                        else -> OSRMRoadManager.MEAN_BY_CAR
+                    },
+                    colorRoad = when (arg.containsKey("roadColor")) {
+                        true -> {
+                            val colors = (arg["roadColor"] as List<Int>)
+                            Color.rgb(colors.first(), colors.last(), colors[1])
+                        }
+                        else -> roadColor
+                    },
+                    roadWidth = when (arg.containsKey("roadWidth")) {
+                        true -> (arg["roadWidth"] as Double).toFloat()
+                        else -> 5f
+                    },
+                    wayPoints = waypoints,
+                    interestPoints = when (arg.containsKey("middlePoints")) {
+                        true -> arg["middlePoints"] as List<HashMap<String, Double>>
+                        false -> emptyList()
+                    }.map { g ->
+                        GeoPoint(g["lat"]!!, g["lon"]!!)
+                    }.toList()
+                )
+            )
+        }
+
+
+        //val showPoiMarker = args["showMarker"] as Boolean
+
+        flutterRoad?.road?.let {
+            map!!.overlays.remove(it)
+        }
+        checkRoadFolderAboveUserOverlay()
+        folderRoad.items.clear()
+        val cachedRoads = map!!.overlays.filterIsInstance<Polyline>().toSet()
+        if (cachedRoads.isNotEmpty()) {
+            map!!.overlays.removeAll(cachedRoads)
+        }
+        map!!.invalidate()
+
+        val resultRoads = emptyList<HashMap<String, Any>>().toMutableList();
+        job = scope?.launch(Default) {
+            withContext(IO) {
+                for (config in listConfigRoad) {
+                    if (roadManager == null)
+                        roadManager = OSRMRoadManager(context, "json/application")
+                    roadManager?.let { manager ->
+                        manager.setMean(config.meanUrl)
+                        var routePointsEncoded = ""
+                        withContext(Main) {
+                            folderMarkers.items.removeAll {
+                                (it is FlutterMarker && config.wayPoints.contains(it.position)) ||
+                                        (it is FlutterMarker && config.interestPoints.contains(it.position))
+                            }
+                            mapSnapShot().removeMarkersFromSnapShot(config.wayPoints)
+                        }
+                        val roadPoints = ArrayList(config.wayPoints)
+                        if (config.interestPoints.isNotEmpty()) {
+                            roadPoints.addAll(1, config.interestPoints)
+                        }
+                        val road = manager.getRoad(roadPoints)
+                        withContext(Main) {
+                            if (road.mRouteHigh.size > 2) {
+                                routePointsEncoded = PolylineEncoder.encode(road.mRouteHigh, 10)
+                                val polyLine = RoadManager.buildRoadOverlay(road)
+                                createRoad(
+                                    polyLine = polyLine,
+                                    colorRoad = config.colorRoad,
+                                    roadWidth = config.roadWidth,
+                                    showPoiMarker = false,
+                                    listInterestPoints = config.interestPoints,
+                                )
+
+                                mapSnapShot().cacheListRoad(
+                                    RoadSnapShot(
+                                        roadPoints = road.mRouteHigh,
+                                        roadColor = config.colorRoad,
+                                        roadWith = config.roadWidth,
+                                        listInterestPoints = config.interestPoints,
+                                        showIcons = false
+                                    )
+                                )
+                                resultRoads.add(HashMap<String, Any>().apply {
+                                    this["duration"] = road.mDuration
+                                    this["distance"] = road.mLength
+                                    this["routePoints"] = routePointsEncoded
+                                })
+                            }
+                        }
+                        delay(100)
+
+                    }
+                }
+
+            }
+            withContext(Main) {
+                map!!.invalidate()
+                result.success(resultRoads.toList())
+            }
+        }
+
+    }
+
+    private fun checkRoadFolderAboveUserOverlay() {
+        if (!map!!.overlays.contains(folderRoad)) {
+            locationNewOverlay?.let { locationOverlay ->
+                val indexOf = map!!.overlays.indexOf(locationOverlay)
+                when (indexOf != -1) {
+                    true -> map!!.overlays.add(indexOf - 1, folderRoad)
+                    false -> map!!.overlays.add(folderRoad)
+                }
+            } ?: map!!.overlays.add(folderRoad)
+        }
+    }
+
     private fun drawRoad(call: MethodCall, result: MethodChannel.Result) {
         val args = call.arguments!! as HashMap<String, Any>
 
@@ -1336,11 +1499,11 @@ class FlutterOsmView(
         flutterRoad?.road?.let {
             map!!.overlays.remove(it)
         }
-        if (!map!!.overlays.contains(folderRoad)) {
-            map!!.overlays.add(folderRoad)
-        } else {
-            folderRoad.items.clear()
-        }
+
+        checkRoadFolderAboveUserOverlay()
+        folderRoad.items.clear()
+
+
         map!!.invalidate()
 
         if (roadManager == null)
@@ -1368,14 +1531,13 @@ class FlutterOsmView(
                     if (road.mRouteHigh.size > 2) {
                         routePointsEncoded = PolylineEncoder.encode(road.mRouteHigh, 10)
                         val polyLine = RoadManager.buildRoadOverlay(road)
-                        createRoad(
+                        flutterRoad = createRoad(
                             polyLine = polyLine,
                             colorRoad = colorRoad,
                             roadWidth = roadWidth,
                             showPoiMarker = showPoiMarker,
                             listInterestPoints = listInterestPoints,
                         )
-
                         mapSnapShot().cacheRoad(
                             RoadSnapShot(
                                 roadPoints = road.mRouteHigh,
@@ -1412,7 +1574,7 @@ class FlutterOsmView(
         showPoiMarker: Boolean,
         listInterestPoints: List<GeoPoint>,
         roadWidth: Float,
-    ) {
+    ): FlutterRoad {
         polyLine.setOnClickListener { _, _, eventPos ->
             methodChannel.invokeMethod("receiveSinglePress", eventPos?.toHashMap())
             true
@@ -1421,13 +1583,13 @@ class FlutterOsmView(
         polyLine.outlinePaint.color = colorRoad ?: Color.GREEN
 
 
-        flutterRoad = FlutterRoad(
+        val flutterRoad = FlutterRoad(
             context,
             map!!,
             interestPoint = if (showPoiMarker) listInterestPoints else emptyList()
         )
 
-        flutterRoad?.let { roadF ->
+        flutterRoad.let { roadF ->
             roadF.markersIcons = customRoadMarkerIcon
             polyLine.outlinePaint.strokeWidth = roadWidth
 
@@ -1446,6 +1608,8 @@ class FlutterOsmView(
 
             folderRoad.items.add(roadF.road!!)
         }
+
+        return flutterRoad
     }
 
     private fun staticPositionIconMaker(call: MethodCall, result: MethodChannel.Result) {