+++ android/src/main/kotlin/hamza/dali/flutter_osm_plugin/models/FlutterInfoWindow.kt
@@ -1,4 +1,4 @@
-package hamza.dali.flutter_osm_plugin
+package hamza.dali.flutter_osm_plugin.models
 
 import android.util.Log
 import android.view.View
@@ -13,7 +13,7 @@ import kotlinx.coroutines.withContext
 import org.osmdroid.util.GeoPoint
 import org.osmdroid.views.MapView
 import org.osmdroid.views.overlay.infowindow.InfoWindow
-
+import hamza.dali.flutter_osm_plugin.utilities.*
 class FlutterInfoWindow(view: View, mapView: MapView?, private val point: GeoPoint) :
     InfoWindow(view, mapView) {
     private var infoView: InfowindowBinding = InfowindowBinding.bind(view)