+++ flutter_osm_interface/lib/src/types/marker.dart
@@ -44,16 +44,13 @@ class AssetMarker {
 
 class MarkerIcon extends StatelessWidget {
   final Icon? icon;
-  @Deprecated("use assetMarker,this will be removed next version")
-  final AssetImage? image;
   final AssetMarker? assetMarker;
 
   const MarkerIcon({
     this.icon,
     this.assetMarker,
-    this.image,
     Key? key,
-  })  : assert(icon != null || (assetMarker != null || image != null)),
+  })  : assert(icon != null || (assetMarker != null)),
         super(key: key);
 
   @override
@@ -61,16 +58,11 @@ class MarkerIcon extends StatelessWidget {
     Widget? child = SizedBox.shrink();
     if (icon != null) {
       child = icon;
-    } else if (assetMarker != null || image != null) {
-      if (image != null) {
-        return Image(
-          image: image!,
-        );
-      }
+    } else if (assetMarker != null) {
       child = Image.asset(
         assetMarker!.image.assetName,
-        scale: assetMarker?.scaleAssetImage,
-        color: assetMarker?.color,
+        scale: assetMarker!.scaleAssetImage,
+        color: assetMarker!.color,
       );
     }